backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
motor==3.5.1
pydantic==2.5.0
httpx==0.25.0
pymongo==4.6.3
certifi>=2024.7.4
python-dotenv>=1.0.1
